Update to the heartbeat packet format to add some additional info

+++ include/typedefs.h
@@ -5,4 +5,4 @@
 #include <FastLED.h>    // For CRGB
 
 typedef std::vector<CRGB> Leds;
-typedef uint8_t Rnd;
+typedef uint16_t Rnd;
